Reverted moving ChainWriter helper methods to utils
