+++ pkg/solana/chainwriter/helpers.go
@@ -3,8 +3,11 @@ package chainwriter
 import (
 	"context"
 	"crypto/sha256"
+	"encoding/binary"
 	"errors"
 	"fmt"
+	"reflect"
+	"strings"
 	"testing"
 
 	"github.com/gagliardetto/solana-go"
@@ -30,8 +33,34 @@ type DataAccount struct {
 	LookupTable          solana.PublicKey
 }
 
+// GetValuesAtLocation parses through nested types and arrays to find all locations of values
+func GetValuesAtLocation(args any, location string) ([][]byte, error) {
+	var vals [][]byte
+	path := strings.Split(location, ".")
+
+	addressList, err := traversePath(args, path)
+	if err != nil {
+		return nil, err
+	}
+	for _, value := range addressList {
+		if byteArray, ok := value.([]byte); ok {
+			vals = append(vals, byteArray)
+		} else if address, ok := value.(solana.PublicKey); ok {
+			vals = append(vals, address.Bytes())
+		} else if num, ok := value.(uint64); ok {
+			buf := make([]byte, 8)
+			binary.LittleEndian.PutUint64(buf, num)
+			vals = append(vals, buf)
+		} else {
+			return nil, fmt.Errorf("invalid value format at path: %s", location)
+		}
+	}
+
+	return vals, nil
+}
+
 func GetDebugIDAtLocation(args any, location string) (string, error) {
-	debugIDList, err := utils.GetValuesAtLocation(args, location)
+	debugIDList, err := GetValuesAtLocation(args, location)
 	if err != nil {
 		return "", err
 	}
@@ -50,6 +79,56 @@ func errorWithDebugID(err error, debugID string) error {
 		return err
 	}
 	return fmt.Errorf("Debug ID: %s: Error: %s", debugID, err)
+}
+
+// traversePath recursively traverses the given structure based on the provided path.
+func traversePath(data any, path []string) ([]any, error) {
+	if len(path) == 0 {
+		return []any{data}, nil
+	}
+
+	var result []any
+
+	val := reflect.ValueOf(data)
+
+	if val.Kind() == reflect.Ptr {
+		val = val.Elem()
+	}
+
+	switch val.Kind() {
+	case reflect.Struct:
+		field := val.FieldByName(path[0])
+		if !field.IsValid() {
+			return nil, errors.New("field not found: " + path[0])
+		}
+		return traversePath(field.Interface(), path[1:])
+
+	case reflect.Slice, reflect.Array:
+		for i := 0; i < val.Len(); i++ {
+			element := val.Index(i).Interface()
+			elements, err := traversePath(element, path)
+			if err == nil {
+				result = append(result, elements...)
+			}
+		}
+		if len(result) > 0 {
+			return result, nil
+		}
+		return nil, errors.New("no matching field found in array")
+
+	case reflect.Map:
+		key := reflect.ValueOf(path[0])
+		value := val.MapIndex(key)
+		if !value.IsValid() {
+			return nil, errors.New("key not found: " + path[0])
+		}
+		return traversePath(value.Interface(), path[1:])
+	default:
+		if len(path) == 1 && val.Kind() == reflect.Slice && val.Type().Elem().Kind() == reflect.Uint8 {
+			return []any{val.Interface()}, nil
+		}
+		return nil, errors.New("unexpected type encountered at path: " + path[0])
+	}
 }
 
 func InitializeDataAccount(
