+++ pkg/solana/utils/utils.go
@@ -3,11 +3,9 @@ package utils
 import (
 	"context"
 	"encoding/binary"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
-	"reflect"
 	"runtime"
 	"strings"
 	"testing"
@@ -218,89 +216,4 @@ func SetupTestValidatorWithAnchorPrograms(t *testing.T, upgradeAuthority string,
 	}
 	rpcURL, wsURL := client.SetupLocalSolNodeWithFlags(t, flags...)
 	return rpcURL, wsURL
-}
-
-func ConvertAnyToPDASeed(value any) []byte {
-	if byteArray, ok := value.([]byte); ok {
-		return byteArray
-	} else if str, ok := value.(string); ok {
-		return []byte(str)
-	} else if address, ok := value.(solana.PublicKey); ok {
-		return address.Bytes()
-	} else if num, ok := value.(uint64); ok {
-		buf := make([]byte, 8)
-		binary.LittleEndian.PutUint64(buf, num)
-		return buf
-	}
-	return nil
-}
-
-// GetValuesAtLocation parses through nested types and arrays to find all locations of values
-func GetValuesAtLocation(args any, location string) ([][]byte, error) {
-	var vals [][]byte
-	path := strings.Split(location, ".")
-
-	addressList, err := traversePath(args, path)
-	if err != nil {
-		return nil, err
-	}
-	for _, value := range addressList {
-		byteArray := ConvertAnyToPDASeed(value)
-		if byteArray == nil {
-			return nil, fmt.Errorf("invalid value format at path: %s", location)
-		}
-		vals = append(vals, byteArray)
-	}
-
-	return vals, nil
-}
-
-// traversePath recursively traverses the given structure based on the provided path.
-func traversePath(data any, path []string) ([]any, error) {
-	if len(path) == 0 {
-		return []any{data}, nil
-	}
-
-	var result []any
-
-	val := reflect.ValueOf(data)
-
-	if val.Kind() == reflect.Ptr {
-		val = val.Elem()
-	}
-
-	switch val.Kind() {
-	case reflect.Struct:
-		field := val.FieldByName(path[0])
-		if !field.IsValid() {
-			return nil, errors.New("field not found: " + path[0])
-		}
-		return traversePath(field.Interface(), path[1:])
-
-	case reflect.Slice, reflect.Array:
-		for i := 0; i < val.Len(); i++ {
-			element := val.Index(i).Interface()
-			elements, err := traversePath(element, path)
-			if err == nil {
-				result = append(result, elements...)
-			}
-		}
-		if len(result) > 0 {
-			return result, nil
-		}
-		return nil, errors.New("no matching field found in array")
-
-	case reflect.Map:
-		key := reflect.ValueOf(path[0])
-		value := val.MapIndex(key)
-		if !value.IsValid() {
-			return nil, errors.New("key not found: " + path[0])
-		}
-		return traversePath(value.Interface(), path[1:])
-	default:
-		if len(path) == 1 && val.Kind() == reflect.Slice && val.Type().Elem().Kind() == reflect.Uint8 {
-			return []any{val.Interface()}, nil
-		}
-		return nil, errors.New("unexpected type encountered at path: " + path[0])
-	}
 }
